requirements.txt
aiohttp==3.9.5
aiohappyeyeballs==2.4.0
aiosignal==1.3.1
aniso8601==9.0.1
anyio==4.6.0
async-timeout==4.0.3
attrs==24.2.0
beautifulsoup4==4.12.3
blis==0.7.11
catalogue==2.0.10
certifi==2024.7.4
cffi==1.17.0
chardet==5.2.0
charset-normalizer==3.3.2
cloudpathlib==0.18.1
confection==0.1.5
cymbolic==2.1.2
dataclasses-json==0.6.7
deprecation==2.1.0
effdet==0.4.1
emmorphlib==1.0.0
fastapi==0.115.0
filelock==3.15.4
frozenlist==1.4.1
fsspec==2024.6.1
greenlet==3.1.0
h11==0.14.0
huggingface-hub==0.24.6
idna==3.7
Jinja2==3.1.4
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.25
langchain-community==0.3.23
langchain-core==0.3.58
langchain-ollama==0.3.2
langchain-text-splitters==0.3.8
langdetect==1.0.9
langserve==0.3.1
langsmith==0.3.42
langgraph==0.0.26
lxml==5.4.0
marshmallow==3.26.1
multidict==6.4.3
mypy_extensions==1.1.0
nest-asyncio==1.6.0
nltk==3.9.1
numexpr==2.10.2
numpy==2.2.5
olefile==0.47
ollama==0.4.8
orjson==3.10.18
packaging==24.2
propcache==0.3.1
psutil==7.0.0
pycparser==2.22
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
pypdf==5.4.0
python-dotenv==1.1.0
python-iso639==2025.2.18
python-magic==0.4.27
python-oxmsg==0.0.2
PyYAML==6.0.2
RapidFuzz==3.13.0
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
six==1.17.0
sniffio==1.3.1
socksio==1.0.0
soupsieve==2.7
SQLAlchemy==2.0.40
starlette==0.46.2
tenacity==9.1.2
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
unstructured==0.17.2
unstructured-client==0.34.0
urllib3==2.4.0
uvicorn==0.34.2
webencodings==0.5.1
wrapt==1.17.2
yarl==1.20.0
zstandard==0.23.0